refactor(panel): extract first-use chain into useFirstUseChain
Amp-Thread-ID: https://ampcode.com/threads/T-019e2394-e4cc-73b9-a95c-23a5a5143a41
