+++ src/renderer/src/panel/PanelApp.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import { computed, onMounted, onUnmounted, ref, watch } from 'vue'
+import { computed, onMounted, onUnmounted, ref } from 'vue'
 import { useI18n } from 'vue-i18n'
 import SettingsModal from '../views/SettingsModal.vue'
 import ProgressModal from '../views/ProgressModal.vue'
@@ -14,15 +14,14 @@ import FirstUseTakeover from '../views/FirstUseTakeover.vue'
 import { useTheme } from '../composables/useTheme'
 import { useSessionStore } from '../stores/sessionStore'
 import { useInstallationStore } from '../stores/installationStore'
-import { useProgressStore } from '../stores/progressStore'
 import { useLauncherPrefs } from '../composables/useLauncherPrefs'
-import { useMigrateAction } from '../composables/useMigrateAction'
 import { useModal } from '../composables/useModal'
 import { useAppUpdatePrompts } from '../composables/useAppUpdatePrompts'
 import { useSendFeedback } from '../composables/useSendFeedback'
 import { useDeepLinkRouter } from '../composables/useDeepLinkRouter'
 import { isFlowPanel, isValidPanel, usePanelOverlays } from './usePanelOverlays'
 import { useChooserHandoff } from './useChooserHandoff'
+import { useFirstUseChain } from './useFirstUseChain'
 import type { Installation } from '../types/ipc'
 
 const { mergeLocaleMessage, locale, t } = useI18n()
@@ -31,30 +30,8 @@ useTheme()
 const params = new URLSearchParams(window.location.search)
 const installationId = params.get('installationId') || ''
 
-/**
- * Set when the first-use takeover's Local branch chains into the
- * new-install Tier 3 takeover. The new-install modal emits `close`
- * after a successful install (the same hook used everywhere); when
- * that fires while this flag is true the host marks `firstUseCompleted`
- * and clears the flag. A Cloud-branch pick marks completion immediately
- * and never sets this flag, so the two paths can't double-fire the
- * pref write.
- */
-const chainingFirstUseToNewInstall = ref(false)
-
-/**
- * Installation id of the Standalone install that the first-use chain
- * (new-install or migration) just kicked off. Captured from the
- * corresponding `show-progress` while a chain flag is set. The
- * progressStore watcher below auto-launches the install once its op
- * finishes successfully, so the user lands on a running ComfyUI as the
- * natural endpoint of first-use without having to click play again.
- */
-const pendingFirstUseAutoLaunchId = ref<string | null>(null)
-
 const sessionStore = useSessionStore()
 const installationStore = useInstallationStore()
-const progressStore = useProgressStore()
 const launcherPrefs = useLauncherPrefs()
 
 const modal = useModal()
@@ -91,17 +68,48 @@ const loadSnapshotRef = ref<InstanceType<typeof LoadSnapshotModal> | null>(null)
 const quickInstallRef = ref<InstanceType<typeof QuickInstallModal> | null>(null)
 const firstUseRef = ref<InstanceType<typeof FirstUseTakeover> | null>(null)
 
-// usePanelOverlays + useChooserHandoff form a small dependency cycle:
-// the overlay's `handleShowProgress` needs to subscribe the chooser
-// host's close-on-instance-started fallback when an install-less
-// host fires a launch-class op; the chooser handoff needs to mount
-// its launches through the overlay's `handleShowProgress` and use
-// `switchPanel` for the missing-launch-action fallback. Break the
-// cycle with a deferred lazy reference: the overlays composable
-// receives a callback that reads `chooserHandoff` at call time, by
-// which point chooserHandoff has been assigned below.
+// The three panel composables form a small dependency cycle:
+//   - useFirstUseChain needs overlay helpers (dismissTakeoverDirect,
+//     switchPanel, handleShowProgress) and chooser launch.
+//   - usePanelOverlays needs the chain's hooks bag + the chooser's
+//     prepareChooserHostHandoff.
+//   - useChooserHandoff needs the overlay's handleShowProgress +
+//     switchPanel.
+// Break the cycle with deferred lazy references: composables that
+// have to be instantiated first take callbacks that read the later-
+// assigned holders at call time (not at construction time). By the
+// time any callback fires (user click, IPC, watcher) all three
+// holders are populated.
+let overlays!: ReturnType<typeof usePanelOverlays>
 let chooserHandoff!: ReturnType<typeof useChooserHandoff>
 
+const firstUseChain = useFirstUseChain({
+  dismissTakeoverDirect: () => overlays.dismissTakeoverDirect(),
+  switchPanel: (panel, entrypoint) => overlays.switchPanel(panel, entrypoint),
+  handleShowProgress: (showOpts) => overlays.handleShowProgress(showOpts),
+  performChooserLaunch: (inst, onMissing) => chooserHandoff.performChooserLaunch(inst, onMissing),
+})
+const {
+  chainingFirstUseToNewInstall,
+  completeFirstUseAndDismiss,
+  handleFirstUseComplete,
+  handleFirstUseChainLocal,
+  handleFirstUseChainMigrate,
+  handleNewInstallTakeoverClose,
+} = firstUseChain
+
+overlays = usePanelOverlays({
+  installationId,
+  installation,
+  progressRef,
+  newInstallRef,
+  trackRef,
+  loadSnapshotRef,
+  quickInstallRef,
+  firstUseRef,
+  prepareChooserHostHandoff: (id) => chooserHandoff.prepareChooserHostHandoff(id),
+  firstUseChain: firstUseChain.hooks,
+})
 const {
   activePanel,
   initialPanel,
@@ -113,43 +121,13 @@ const {
   openFirstUseTakeover,
   dismissTakeoverDirect,
   switchPanel,
-} = usePanelOverlays({
-  installationId,
-  installation,
-  progressRef,
-  newInstallRef,
-  trackRef,
-  loadSnapshotRef,
-  quickInstallRef,
-  firstUseRef,
-  prepareChooserHostHandoff: (id) => chooserHandoff.prepareChooserHostHandoff(id),
-  firstUseChain: {
-    shouldForceTakeover: () => chainingFirstUseToNewInstall.value,
-    onShowProgress: (showOpts) => {
-      // Capture the operation's installation id when a first-use chain
-      // is in flight (new-install or migrate). The progressStore
-      // watcher below auto-launches the resulting install once the op
-      // finishes successfully. New-install ops carry the new install's
-      // id directly; migrate ops carry the Legacy Desktop install's id
-      // and the watcher resolves the resulting Standalone install
-      // after the op finishes. Only the first chained op captures the
-      // id — subsequent show-progress calls leave it untouched.
-      if (chainingFirstUseToNewInstall.value && pendingFirstUseAutoLaunchId.value === null) {
-        pendingFirstUseAutoLaunchId.value = showOpts.installationId
-        // Flip the persisted gate now so the takeover doesn't re-run
-        // on the next launch — the overlay handoff doesn't go through
-        // NewInstallModal's close emit.
-        void launcherPrefs.markFirstUseCompleted()
-      }
-    },
-  },
-})
+} = overlays
 
 chooserHandoff = useChooserHandoff({
   showProgress: handleShowProgress,
   switchPanel,
 })
-const { performChooserLaunch, handleChooserPick, handleChooserShowNewInstall } = chooserHandoff
+const { handleChooserPick, handleChooserShowNewInstall } = chooserHandoff
 
 let unsubPanel: (() => void) | null = null
 let unsubLocale: (() => void) | null = null
@@ -176,228 +154,6 @@ async function loadLocale(): Promise<void> {
   mergeLocaleMessage('en', messages)
   locale.value = 'en'
 }
-
-/** Shared completion helper. The Cloud-branch pick
- *  (`handleFirstUseComplete`), the file-menu Skip Onboarding entry
- *  (`onFirstUseSkip` listener) and the new-install chain close
- *  (`handleNewInstallTakeoverClose`) all run the same
- *  `markFirstUseCompleted` → dismiss sequence; extracting the pair
- *  keeps them in sync if the gate flip ever needs extra state cleanup. */
-async function completeFirstUseAndDismiss(): Promise<void> {
-  // Clear chain state so the auto-launch watcher doesn't fire after a
-  // Skip Onboarding triggered mid-chain (the user wants OUT of
-  // onboarding, not to land on a freshly-installed Comfy).
-  chainingFirstUseToNewInstall.value = false
-  pendingFirstUseAutoLaunchId.value = null
-  await launcherPrefs.markFirstUseCompleted()
-  // dismissTakeoverDirect pushes `'none'` only when the overlay is
-  // the first-use takeover itself; chain dismiss paths can have a
-  // new-install / progress takeover in the slot, so push it
-  // explicitly to keep the file-menu builder in steady state.
-  window.api.setFirstUseMode('none')
-  dismissTakeoverDirect()
-}
-
-/** First-use takeover Cloud-branch pick (`complete-cloud` emit) — the
- *  user explicitly picked Cloud at the cloud-vs-local fork. Mark
- *  completion, close the takeover, and auto-launch the always-present
- *  Cloud install so they reach a running ComfyUI as the natural
- *  endpoint of first-use without having to click play again. The
- *  launch goes through the same `useListAction` pipeline the chooser
- *  uses (close-host-window-on-instance-started, etc.). If the cloud
- *  install can't be found we still mark complete and close the
- *  takeover — the chooser body underneath is the fallback.
- *
- *  The returning-user `complete-skip` emit is wired directly to
- *  `completeFirstUseAndDismiss` instead — those users never picked
- *  Cloud (the fork was suppressed), so auto-launching it would hijack
- *  their existing local install. */
-async function handleFirstUseComplete(): Promise<void> {
-  chainingFirstUseToNewInstall.value = false
-  // Mark completion but DON'T dismiss the takeover yet — dismissing
-  // first would expose the dashboard underneath while the launch
-  // action races to mount its own progress overlay. Cloud's launch
-  // action has `showProgress: true`, so the launch goes through
-  // `handleShowProgress` → `openOverlay({ kind: 'takeover',
-  // component: 'update' })` which silently swaps the first-use
-  // takeover for the connect-progress takeover (Tier 3 → Tier 3).
-  await launcherPrefs.markFirstUseCompleted()
-  window.api.setFirstUseMode('none')
-  // Find the auto-seeded Cloud install. The store may not be hydrated
-  // yet on first-launch, so we fall back to a fresh fetch via main.
-  let cloud = installationStore.installations.find((i) => i.sourceCategory === 'cloud') ?? null
-  if (!cloud) {
-    try {
-      const all = await window.api.getInstallations()
-      cloud = all.find((i) => i.sourceCategory === 'cloud') ?? null
-    } catch {}
-  }
-  if (cloud) {
-    // If the cloud install has no resolvable launch action (defensive
-    // fallback — production cloud sources always provide one), dismiss
-    // the takeover so the user lands on the chooser body. The happy
-    // path leaves the takeover up; the launch action's `showProgress:
-    // true` swaps it for a connect-progress takeover via
-    // `handleShowProgress` (Tier 3 → Tier 3 silent).
-    await performChooserLaunch(cloud, dismissTakeoverDirect)
-    return
-  }
-  // No cloud install to launch into — fall back to dismissing the
-  // takeover so the user lands on the chooser body underneath.
-  dismissTakeoverDirect()
-}
-
-/** First-use takeover Local-branch pick — chain into the new-install
- *  Tier 3 takeover. The Tier 3 → Tier 3 swap is silent in
- *  `useOverlay`, so the first-use takeover unmounts as the new-install
- *  takeover mounts. The completion flip is deferred to the new-install
- *  close path (see `handleNewInstallTakeoverClose`). */
-async function handleFirstUseChainLocal(): Promise<void> {
-  chainingFirstUseToNewInstall.value = true
-  pendingFirstUseAutoLaunchId.value = null
-  await switchPanel('new-install', 'first_use')
-  // FirstUseTakeover.onUnmounted just pushed `'none'` as the chain
-  // swap unmounted it. Re-assert `'post-consent'` so the file-menu
-  // builder keeps the chain locked down to Skip Onboarding while
-  // the new-install / install-progress takeover is up.
-  window.api.setFirstUseMode('post-consent')
-}
-
-/** First-use takeover migrate-branch pick — runs the migrate-to-
- *  standalone action against the auto-tracked Legacy Desktop install.
- *  Same shape as the chain-local path: the migration progress op
- *  flows through `handleShowProgress` (Tier 2 progress modal),
- *  capturing `pendingFirstUseAutoLaunchId` for the resulting
- *  Standalone install along the way. The progressStore watcher below
- *  auto-launches once the op finishes successfully. */
-const { confirmMigration } = useMigrateAction()
-async function handleFirstUseChainMigrate(): Promise<void> {
-  let legacy = installationStore.installations.find((i) => i.sourceId === 'desktop') ?? null
-  if (!legacy) {
-    try {
-      const all = await window.api.getInstallations()
-      legacy = all.find((i) => i.sourceId === 'desktop') ?? null
-    } catch {}
-  }
-  if (!legacy) {
-    // Detection drift — main flagged hasLegacyDesktop=true but the
-    // install is gone now. Bail to chain-local so the user still gets
-    // to the new-install Standalone path.
-    handleFirstUseChainLocal()
-    return
-  }
-  // confirmMigration shows its own modal (variant pick / preview /
-  // confirm); a `null` return means user cancelled, in which case we
-  // leave the takeover mounted on the localBranch step (no state
-  // change).
-  const result = await confirmMigration(legacy)
-  if (!result) return
-
-  // Pre-mark the chain so the new install kicked off by migration
-  // gets captured as the auto-launch target. The migration emits
-  // installations-changed when the new Standalone install is added,
-  // and `handleShowProgress` carries the `installationId` through —
-  // we record it there.
-  chainingFirstUseToNewInstall.value = true
-  pendingFirstUseAutoLaunchId.value = null
-  // Dismiss the takeover before kicking off the migration so the
-  // Tier 2 progress modal isn't blocked by the takeover overlay.
-  dismissTakeoverDirect()
-  await handleShowProgress({
-    installationId: legacy.id,
-    title: `Migrating — ${legacy.name}`,
-    apiCall: () => window.api.runAction(legacy!.id, 'migrate-to-standalone', result),
-    cancellable: true,
-  })
-  // dismissTakeoverDirect pushed `'none'` as it cleared the first-use
-  // overlay; re-assert `'post-consent'` so the file-menu builder
-  // keeps the chain locked down to Skip Onboarding for the duration
-  // of the migration progress + auto-launch.
-  window.api.setFirstUseMode('post-consent')
-}
-
-/** Wrapper around `closeOverlay` for the new-install takeover branch
- *  that also flips `firstUseCompleted` when the close arrives at the
- *  end of a first-use → Local chain. The new-install modal emits
- *  `close` after a successful install AND on user-cancel (✕); both
- *  cases count as "the user got past the cloud-or-local pick". */
-async function handleNewInstallTakeoverClose(): Promise<void> {
-  if (chainingFirstUseToNewInstall.value) {
-    await launcherPrefs.markFirstUseCompleted()
-    // The chain pushed `'post-consent'` to keep Skip Onboarding the
-    // only file-menu entry while the new-install takeover was up.
-    // Clear it here — whatever follows (dismiss back to chooser, or
-    // an in-flight progress / launch overlay swap) is post-onboarding.
-    window.api.setFirstUseMode('none')
-    // Don't clear `chainingFirstUseToNewInstall` yet — the auto-launch
-    // watcher uses it together with `pendingFirstUseAutoLaunchId` to
-    // decide whether to fire. The watcher clears both after launch.
-  }
-  // Only dismiss when the new-install takeover is still in the slot.
-  // The happy-path install handoff in NewInstallModal swaps the
-  // overlay to a progress takeover via @show-progress without first
-  // emitting `close`, but `@navigate-list` still routes here for the
-  // skipInstall branch — and dismissing then would clear an unrelated
-  // overlay if anything else has claimed the slot in between.
-  if (
-    currentOverlay.value?.kind === 'takeover'
-    && currentOverlay.value.component === 'new-install'
-  ) {
-    dismissTakeoverDirect()
-  }
-}
-
-/** Watch progressStore for the new-install / migration op finishing
- *  and auto-launch the resulting Standalone install. The chain flag
- *  (`chainingFirstUseToNewInstall`) gates the watcher so we only
- *  auto-launch when the op was actually driven by the first-use
- *  chain. The captured id is set inside `usePanelOverlays`'
- *  `firstUseChain.onShowProgress` hook at the moment the chained op
- *  begins. */
-watch(
-  () => {
-    const id = pendingFirstUseAutoLaunchId.value
-    if (!id) return null
-    const op = progressStore.operations.get(id)
-    return op && op.finished ? op : null
-  },
-  async (op) => {
-    if (!op) return
-    if (!chainingFirstUseToNewInstall.value) return
-    const id = pendingFirstUseAutoLaunchId.value
-    chainingFirstUseToNewInstall.value = false
-    pendingFirstUseAutoLaunchId.value = null
-    // Chain is done (success or failure) — drop the file-menu lock.
-    // The launch path that follows (when the op succeeded) replaces
-    // the install-progress takeover with its own connect-progress
-    // takeover; either way the user is past onboarding now.
-    window.api.setFirstUseMode('none')
-    if (!id) return
-    if (op.cancelRequested || op.error || !op.result?.ok) return
-    // The migrate-to-standalone op runs against the Legacy Desktop
-    // install but produces a fresh Standalone install — wait for the
-    // store to reflect the new install, then launch the most-recently-
-    // created non-cloud, non-legacy local install (the migration's
-    // result). For new-install ops, the captured id is the new install's
-    // id directly so this branch resolves immediately.
-    let inst = installationStore.installations.find((i) => i.id === id) ?? null
-    if (!inst || inst.sourceId === 'desktop') {
-      try {
-        await installationStore.fetchInstallations()
-      } catch {}
-      inst = installationStore.installations.find(
-        (i) => (i as unknown as { copiedFrom?: string }).copiedFrom === id,
-      ) ?? installationStore.installations
-        .filter((i) => i.sourceCategory === 'local')
-        .sort((a, b) => Date.parse(String(b.createdAt ?? '')) - Date.parse(String(a.createdAt ?? '')))[0]
-        ?? null
-    }
-    if (inst) {
-      void performChooserLaunch(inst)
-    }
-  },
-  { deep: false },
-)
 
 function handleUpdateInstallation(inst: Installation): void {
   // Optimistic local update for snappier UX while the broadcast-driven
